feat(ext-openai): move createOpenAI*Runtime factories out of core
Moves all OpenAI-specific runtime code (~700 lines) from
src/provider/runtime-loader.ts into extensions/ext-openai/src/openai-provider.ts.
The core now throws a clear install hint when ext-openai is absent.
Error classes remain in runtime-loader.ts and are re-exported from the
shared barrel so extension tests can import them without touching core internals.
