+++ src/provider/shared/index.ts
@@ -40,6 +40,11 @@ export {
   isNumberArray,
   mergeUsage,
   parseRetryAfterMs,
+  ProviderError,
+  ProviderOverloadedError,
+  ProviderQuotaError,
+  ProviderRateLimitError,
+  ProviderRequestError,
   readProviderOptions,
   readRecord,
   readTextParts,
